Revisions

+++ TouristAgency/Controllers/ClientsController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using TouristAgency.Data;
+using TouristAgency.EntityServices;
 using TouristAgency.Models;
 
 namespace TouristAgency.Controllers
@@ -13,21 +14,60 @@ namespace TouristAgency.Controllers
     public class ClientsController : Controller
     {
         private readonly TouristAgencyContext _context;
+        private readonly ClientService _service;
+        private readonly int _pageSize;
 
         public ClientsController(TouristAgencyContext context)
         {
             _context = context;
+            _service = new ClientService();
+            _pageSize = 5;
         }
 
         // GET: Clients
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string selectedLastName, int? page, ClientService.SortState? sortState)
         {
-            return View(await _context.Clients.ToListAsync());
+            if (!User.IsInRole(Areas.Identity.Roles.User) && !User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return Redirect("~/Identity/Account/Login");
+            }
+
+            bool isFromFilter = HttpContext.Request.Query["isFromFilter"] == "true";
+
+            _service.GetSortPagingCookiesForUserIfNull(Request.Cookies, User.Identity.Name, isFromFilter,
+                ref page, ref sortState);
+            _service.GetFilterCookiesForUserIfNull(Request.Cookies, User.Identity.Name, isFromFilter,
+                ref selectedLastName);
+            _service.SetDefaultValuesIfNull(ref selectedLastName, ref page, ref sortState);
+            _service.SetCookies(Response.Cookies, User.Identity.Name, selectedLastName, page, sortState);
+
+            var clients = _context.Clients.AsQueryable();
+
+            clients = _service.Filter(clients, selectedLastName);
+
+            var count = await clients.CountAsync();
+
+            clients = _service.Sort(clients, (ClientService.SortState)sortState);
+            clients = _service.Paging(clients, isFromFilter, (int)page, _pageSize);
+
+            ViewModels.Client.IndexClientViewModel model = new ViewModels.Client.IndexClientViewModel
+            {
+                Clients = await clients.ToListAsync(),
+                PageViewModel = new ViewModels.PageViewModel(count, (int)page, _pageSize),
+                FilterClientViewModel = new ViewModels.Client.FilterClientViewModel(selectedLastName),
+                SortClientViewModel = new ViewModels.Client.SortClientViewModel((ClientService.SortState)sortState),
+            };
+
+            return View(model);
         }
 
         // GET: Clients/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!User.IsInRole(Areas.Identity.Roles.User) && User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return Redirect("~/Identity/Account/Login");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -46,6 +86,10 @@ namespace TouristAgency.Controllers
         // GET: Clients/Create
         public IActionResult Create()
         {
+            if (!User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return RedirectToAction("Index", "Clients");
+            }
             return View();
         }
 
@@ -56,6 +100,10 @@ namespace TouristAgency.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,LastName,FirstName,MiddleName,BirthDate,Gender,Address,Phone,PassportData,Discount")] Client client)
         {
+            if (!User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return RedirectToAction("Index", "Clients");
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(client);
@@ -68,6 +116,10 @@ namespace TouristAgency.Controllers
         // GET: Clients/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return RedirectToAction("Index", "Clients");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -88,6 +140,10 @@ namespace TouristAgency.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,LastName,FirstName,MiddleName,BirthDate,Gender,Address,Phone,PassportData,Discount")] Client client)
         {
+            if (!User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return RedirectToAction("Index", "Clients");
+            }
             if (id != client.Id)
             {
                 return NotFound();
@@ -119,6 +175,10 @@ namespace TouristAgency.Controllers
         // GET: Clients/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return RedirectToAction("Index", "Clients");
+            }
             if (id == null)
             {
                 return NotFound();
@@ -139,6 +199,10 @@ namespace TouristAgency.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!User.IsInRole(Areas.Identity.Roles.Admin))
+            {
+                return RedirectToAction("Index", "Clients");
+            }
             var client = await _context.Clients.FindAsync(id);
             _context.Clients.Remove(client);
             await _context.SaveChangesAsync();
